Clean up code

+++ TheaterSuggestions/Java/SeatsSuggestions/src/main/java/com/octo/technology/SeatsSuggestions/DeepModel/OfferSeatingPlacesNearerTheMiddleOfTheRow.java
@@ -38,7 +38,7 @@ public  class OfferSeatingPlacesNearerTheMiddleOfTheRow
                 .skip(middleOfTheRow)
                 .map(s -> new SeatWithDistance(
                         s,
-                        (int)(s.number() -  computeTheMiddleOfTheRowForRightSide(middleOfTheRow))
+                        s.number() -  computeTheMiddleOfTheRowForRightSide(middleOfTheRow)
                 ))
                 .collect(Collectors.toList());
     }
